lean, fighting shape
