+++ javascript/transports/action_cable.js
@@ -1,8 +1,26 @@
+import { getConsumer } from '../consumer'
+import { performOperations } from '../reflexes'
 import { emitEvent } from '../utils'
 
 let consumer
 let params
 let subscriptionActive
+
+const createSubscription = controller => {
+  consumer = consumer || getConsumer()
+  const { channel } = controller.StimulusReflex
+  const subscription = { channel, ...params }
+  const identifier = JSON.stringify(subscription)
+
+  controller.StimulusReflex.subscription =
+    consumer.subscriptions.findAll(identifier)[0] ||
+    consumer.subscriptions.create(subscription, {
+      received: performOperations,
+      connected,
+      rejected,
+      disconnected
+    })
+}
 
 const connected = () => {
   subscriptionActive = true
@@ -26,6 +44,7 @@ export default {
   get subscriptionActive () {
     return subscriptionActive
   },
+  createSubscription,
   connected,
   rejected,
   disconnected
